@@ -10,10 +10,10 @@ import (
 	"strconv"
 	"sync"
 
-    "go.containerssh.io/libcontainerssh/config"
-    "go.containerssh.io/libcontainerssh/internal/metrics"
-    "go.containerssh.io/libcontainerssh/log"
-    "go.containerssh.io/libcontainerssh/message"
+	"go.containerssh.io/libcontainerssh/config"
+	"go.containerssh.io/libcontainerssh/internal/metrics"
+	"go.containerssh.io/libcontainerssh/log"
+	"go.containerssh.io/libcontainerssh/message"
 	"k8s.io/client-go/tools/remotecommand"
 	"k8s.io/client-go/util/exec"
 )
@@ -298,7 +298,7 @@ func (k *kubernetesExecutionImpl) handleStream(
 		tty = k.tty
 	}
 	k.backendRequestsMetric.Increment()
-	err := k.exec.Stream(
+	err := k.exec.StreamWithContext(context.Background(),
 		remotecommand.StreamOptions{
 			Stdin:             stdin,
 			Stdout:            stdout,
